# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CortexCloud API",
    "version": "1.0.0",
    "description": "Production-ready OpenAI-compatible AI gateway with x402 payment-gated routes. Agents pay per call in USDC on Base (eip155:8453) via the x402 protocol \u2014 no API key or subscription required.",
    "contact": {
      "email": "ampwerajonathan50@gmail.com"
    },
...
  },
  "servers": [
    {
      "url": "https://api.cortexcloud.org"
    }
  ],
  "paths": {
    "/x402/v1/chat/completions": {
      "post": {
        "operationId": "chat_completions",
        "summary": "OpenAI-compatible chat completion (x402 payment-gated)",
        "description": "Send an OpenAI-style chat completion request. Returns HTTP 402 with an x402 PaymentRequirements challenge (USDC on Base) unless a valid X-PAYMENT header is supplied.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.005"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "type": "string"
                  },
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  },
                  "stream": {
                    "type": "boolean"
                  },
                  "temperature": {
                    "type": "number"
                  },
                  "max_tokens": {
                    "type": "integer"
                  }
                },
                "required": [
                  "model",
                  "messages"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "id": "chatcmpl-abc",
                    "object": "chat.completion",
                    "choices": [
                      {
                        "index": 0,
                        "message": {
                          "role": "assistant",
                          "content": "..."
                        },
                        "finish_reason": "stop"
                      }
                    ],
                    "usage": {
                      "prompt_tokens": 10,
                      "completion_tokens": 5,
                      "total_tokens": 15
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
    "/x402/v1/responses": {
      "post": {
        "operationId": "responses",
        "summary": "OpenAI Responses API completion (x402 payment-gated)",
        "description": "OpenAI Responses API endpoint. Returns HTTP 402 with an x402 PaymentRequirements challenge (USDC on Base) unless a valid X-PAYMENT header is supplied.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.005"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "type": "string"
                  },
                  "input": {
                    "type": "string"
                  },
                  "stream": {
                    "type": "boolean"
                  },
                  "temperature": {
                    "type": "number"
                  },
                  "max_output_tokens": {
                    "type": "integer"
                  }
                },
                "required": [
                  "model",
                  "input"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "id": "resp_abc",
                    "object": "response",
                    "output": [
                      {
                        "type": "message",
                        "role": "assistant",
                        "content": [
                          {
                            "type": "output_text",
                            "text": "..."
                          }
                        ]
                      }
                    ]
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
    "/x402/v1/embeddings": {
      "post": {
        "operationId": "embeddings",
        "summary": "OpenAI-compatible text embeddings (x402 payment-gated)",
        "description": "Generate text embeddings. Returns HTTP 402 with an x402 PaymentRequirements challenge (USDC on Base) unless a valid X-PAYMENT header is supplied.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.005"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "input": {
                    "type": "string"
                  },
                  "model": {
                    "type": "string"
                  }
                },
                "required": [
                  "input"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "id": "emb_abc",
                    "object": "list",
                    "data": [
                      {
                        "object": "embedding",
                        "embedding": [
                          0.01,
                          -0.02
                        ],
                        "index": 0
                      }
                    ],
                    "usage": {
                      "prompt_tokens": 8,
                      "total_tokens": 8
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
    "/x402/v1/data/prices": {
      "get": {
        "operationId": "data_prices",
        "summary": "Spot token prices (x402 payment-gated, CoinGecko)",
        "description": "Real-time spot prices from CoinGecko. Returns HTTP 402 with an x402 challenge unless a valid X-PAYMENT header is supplied. No caller-side API key required.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.005"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated CoinGecko coin ids, e.g. bitcoin,ethereum"
          },
          {
            "name": "vs",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated vs currencies, e.g. usd,eur"
          }
        ],
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "bitcoin": {
                      "usd": 67000.0,
                      "eur": 62000.0
                    }
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
    "/x402/v1/data/coins/search": {
      "get": {
        "operationId": "data_coin_search",
        "summary": "Search coins (x402 payment-gated, CoinGecko)",
        "description": "Search CoinGecko coins by name/symbol. Returns HTTP 402 with an x402 challenge unless a valid X-PAYMENT header is supplied.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.005"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Coin name or symbol to search"
          }
        ],
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "coins": [
                      {
                        "id": "bitcoin",
                        "symbol": "btc",
                        "name": "Bitcoin"
                      }
                    ]
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
    "/x402/v1/data/dex/search": {
      "get": {
        "operationId": "data_dex_search",
        "summary": "Search DEX pairs (x402 payment-gated, DEXScreener)",
        "description": "Search DEX pairs/liquidity on DEXScreener by token or address. Returns HTTP 402 with an x402 challenge unless a valid X-PAYMENT header is supplied.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.005"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token symbol or address to search, e.g. WETH"
          }
        ],
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "pairs": [
                      {
                        "chainId": "ethereum",
                        "dexId": "uniswap-v2",
                        "baseToken": {
                          "symbol": "WETH"
                        },
                        "quoteToken": {
                          "symbol": "USDC"
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
    "/x402/v1/data/dex/pairs": {
      "get": {
        "operationId": "data_dex_pairs",
        "summary": "Top DEX pairs for a token (x402 payment-gated, DEXScreener)",
        "description": "Top DEX pairs for a token on a given chain via DEXScreener. Returns HTTP 402 with an x402 challenge unless a valid X-PAYMENT header is supplied.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.005"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "chain",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Chain id, e.g. ethereum, base, solana"
          },
          {
            "name": "pair",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token address or pair address"
          }
        ],
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "pairs": [
                      {
                        "chainId": "ethereum",
                        "dexId": "uniswap-v2",
                        "priceUsd": "3500.12"
                      }
                    ]
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
    "/x402/v1/data/base/balance": {
      "get": {
        "operationId": "base_balance",
        "summary": "Native ETH balance on Base (x402 payment-gated, public RPC)",
        "description": "Native ETH balance of a Base address via public Base JSON-RPC. Returns HTTP 402 with an x402 challenge unless a valid X-PAYMENT header is supplied. Keyless.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.005"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Base wallet address, e.g. 0x1234...abcd"
          }
        ],
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "address": "0x...",
                    "balance": "1234567890123456789",
                    "network": "base"
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
    "/x402/v1/data/base/token-balance": {
      "get": {
        "operationId": "base_token_balance",
        "summary": "ERC-20 token balance on Base (x402 payment-gated, public RPC)",
        "description": "ERC-20 token balance for a Base address via public Base JSON-RPC (balanceOf). token may be an address or symbol (usdc, weth, dai, usdt). Returns HTTP 402 with an x402 challenge unless a valid X-PAYMENT header is supplied. Keyless.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.005"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Base wallet address"
          },
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token address OR symbol (usdc, weth, dai, usdt)"
          }
        ],
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "address": "0x...",
                    "token": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                    "network": "base",
                    "raw": "5000000",
                    "note": "balance is in token base units (USDC has 6 decimals)"
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
    "/x402/v1/data/base/nonce": {
      "get": {
        "operationId": "base_nonce",
        "summary": "Account nonce on Base (x402 payment-gated, public RPC)",
        "description": "Transaction count (nonce) for a Base address via public Base JSON-RPC. Returns HTTP 402 with an x402 challenge unless a valid X-PAYMENT header is supplied. Keyless.",
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.005"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Base wallet address"
          }
        ],
        "responses": {
          "200": {
            "description": "Result (after valid x402 payment)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "address": "0x...",
                    "nonce": 7
                  }
                }
              }
            }
          },
          "402": {
            "description": "Payment Required \u2014 x402 PaymentRequirements challenge"
          }
        }
      }
    },
...
  }
}